Create universal build and streamline build
Creates a universal build for GLIBC, musl, macOS and the various OpenSSL
permutations that can found on those platforms. This is to streamline
the install process so that we don't need to continue creating new
libraries everytime a new distribution release is made.

Adds an OpenSSL 3.x build for all 3 platforms but this is mostly
untested as it's still an alpha and no platform has moved to this new
version yet. This should hopefully just work as it becomes more
prevalent but if not the build framework is in place.

Adds support for a dynamic OpenSSL source on macOS to support installs
from both brew and port. The Install-WSMan cmdlet now detects which one
is installed and sets up various symlinks to better match the platform
it is installing on.

Finally the build process has been tweaked to hopefully support
universal builds (x86_64 and arm64) on macOS which can be enabled once
Azure Pipelines adds the Big Sur target which contains the SDK required
to do this.